+++ .github/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Update <small>_ December 2022</small>
 
+- feat: stickies for forum posts which get posted on forum post create (19/12/2022)
 - refactor: outdated information + links in commands (08/12/2022)
 - fix: fix latlong image typo and prime meridian alignment (04/12/2022)
 - refactor: update wb command for change in Stable (03/12/2022)